+++ shared/SalesforceSDKCore/SalesforceSDKCore/Classes/SmartStore/SFQuerySpec.m
@@ -336,11 +336,11 @@ NSString * const kQuerySpecParamSmartSql = @"smartSql";
 
 - (NSString*)description {
     if (self.queryType == kSFSoupQueryTypeSmart) {
-        return [NSString stringWithFormat:@"<SFSoupQuerySpec: %p> { \n  queryType:\"%d\" \n smartSql:\"%@\" \n pageSize: %d}", self,self.queryType, self.smartSql,self.pageSize];
+        return [NSString stringWithFormat:@"<SFSoupQuerySpec: %p> { \n  queryType:\"%d\" \n smartSql:\"%@\" \n pageSize: %lu}", self,self.queryType, self.smartSql,(unsigned long)self.pageSize];
     }
     else {
-        return [NSString stringWithFormat:@"<SFSoupQuerySpec: %p> { \n  queryType:\"%d\" \n soupName:\"%@\" \n smartSql:\"%@\" \n path:\"%@\" \n beginKey:\"%@\" \n endKey:\"%@\" \n  order:%d \n pageSize: %d}",
-                self,self.queryType, self.soupName, self.smartSql, self.path,self.beginKey,self.endKey,self.order,self.pageSize];
+        return [NSString stringWithFormat:@"<SFSoupQuerySpec: %p> { \n  queryType:\"%d\" \n soupName:\"%@\" \n smartSql:\"%@\" \n path:\"%@\" \n beginKey:\"%@\" \n endKey:\"%@\" \n  order:%d \n pageSize: %lu}",
+                self,self.queryType, self.soupName, self.smartSql, self.path,self.beginKey,self.endKey,self.order,(unsigned long)self.pageSize];
     }
 }
